[Refactor] #4025 rooms-pit-nest.cpp/h からrooms-pit.cpp/h を分離させた
