+++ src/room/rooms-pit.h
@@ -1,0 +1,6 @@
+#pragma once
+
+class PlayerType;
+struct dun_data_type;
+bool build_type6(PlayerType *player_ptr, dun_data_type *dd_ptr);
+bool build_type13(PlayerType *player_ptr, dun_data_type *dd_ptr);
